seeders, migrations for newly-created project-related models

+++ routes/api.php
@@ -2,10 +2,11 @@
 
 use App\Http\Resources\UserResource;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
-    return new UserResource($request->user());
+    return new UserResource(Auth::user()->load('roles'));
 });
 
 //Route::group((array)'/admin', function () {
@@ -28,13 +29,13 @@ Route::prefix('/admin')->group(function () {
 Route::prefix('/file')->group(function () {
     Route::controller(\App\Http\Controllers\Project\ProjectFileController::class)->group(function () {
         Route::post('/upload', 'upload');
-    });
+    })->middleware(['auth:sanctum']); // there needs to be a middleware that controls access to file upload
 });
 
 Route::controller(App\Http\Controllers\Project\ProjectController::class)->group(function () {
     Route::get('/projects', 'index');
     Route::get('/projects/{id}', 'show');
-    Route::post('/projects', 'store');
-    Route::put('/projects/{id}', 'update');
-    Route::delete('/projects/{id}', 'destroy');
+    Route::post('/projects', 'store')->middleware(['auth:sanctum']); // there needs to be a middleware that controls access to proj creation
+    Route::put('/projects/{id}', 'update')->middleware(['auth:sanctum']);
+    Route::delete('/projects/{id}', 'destroy')->middleware(['auth:sanctum']);
 });
